+++ api/pc/image.go
@@ -12,9 +12,9 @@ type ImageListReq struct {
 }
 type ImageListRes struct {
 	List  []*model.ImageListItem
-	Page  int   `json:"page" description:"分页码"`
-	Size  int   `json:"size" description:"分页数量"`
-	Total int64 `json:"total" description:"数据总数"`
+	Page  int `json:"page" description:"分页码"`
+	Size  int `json:"size" description:"分页数量"`
+	Total int `json:"total" description:"数据总数"`
 }
 
 type ImageDetailReq struct {
