+++ api/pc/search.go
@@ -13,7 +13,7 @@ type SearchReq struct {
 }
 type SearchRes struct {
 	List  []*model.ArticleListItem
-	Page  int   `json:"page" description:"分页码"`
-	Size  int   `json:"size" description:"分页数量"`
-	Total int64 `json:"total" description:"数据总数"`
+	Page  int `json:"page" description:"分页码"`
+	Size  int `json:"size" description:"分页数量"`
+	Total int `json:"total" description:"数据总数"`
 }
