+++ internal/service/page_info.go
@@ -11,8 +11,8 @@ import (
 
 type (
 	IPageInfo interface {
-		LayUiPageInfo(ctx context.Context, total int64, size int) string
-		PcPageInfo(ctx context.Context, total int64, size int) string
+		LayUiPageInfo(ctx context.Context, total int, size int) string
+		PcPageInfo(ctx context.Context, total int, size int) string
 	}
 )
